versatilepb: Readjust memory layout

+++ utils/string.c
@@ -17,7 +17,7 @@ void *memset(void *s, int c, size_t n)
 {
   uint8_t *ptr = (uint8_t *)s;
 
-  for (int i = 0; i < n; i++)
+  for (size_t i = 0; i < n; i++)
     *(ptr + i) = c;
 
   return s;
@@ -36,7 +36,7 @@ void *memset(void *s, int c, size_t n)
  */
 void *memcpy(void *dest, const void *src, size_t len)
 {
-  for (int i = 0; i < len; i++)
+  for (size_t i = 0; i < len; i++)
   {
     *((uint8_t *)dest + i) = *((uint8_t *)src + i);
   }
